Use ctrl keys in window mappings where possible.
These are slightly easier to type, so they should be preferred.
Also, for any mapping of the form «C-w»«C-?», the mapping «C-w»? will
also work, but the vice versa is not true. So, using ctrl keys where
possible makes it clearer when it is *not* possible to use them (for
example, «C-w»«C-c» does *not* close a window, and «C-w»«C-s» does *not*
split a window).